% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mafft_map.R
\name{read_mafft_map}
\alias{read_mafft_map}
\title{read_mafft_map}
\usage{
read_mafft_map(mafft_map_path)
}
\arguments{
\item{mafft_map_path}{}
}
\value{
A tibble.
}
\description{
Read in a file output by mafft when the --mapout flag is used.
}
\examples{
read_mafft_map("tests/testthat/O24426_CHLRE.fasta.map")
}
